+++ projects/egeo/src/lib/st-dynamic-table/st-dynamic-table.model.ts
@@ -3,6 +3,13 @@ import { StTableHeader } from '../';
 
 export interface StDynamicTableHeader extends StTableHeader {
    reference: string;
+   fk?: StDynamicTableFk;
+   group?: string;
+}
+
+export interface StDynamicTableFkEvent {
+   fk?: StDynamicTableFk;
+   value: string | number;
 }
 
 export interface StDynamicTableUISpecification {
@@ -10,10 +17,7 @@ export interface StDynamicTableUISpecification {
    styles?: {
       [key: string]: string;
    };
-   fk?: {
-      table: string;
-      field: string;
-   };
+   fk?: StDynamicTableFk;
    group_field?: {
       name: string;
    };
@@ -23,4 +27,9 @@ export interface StDynamicTableUISpecification {
 
 export interface StDynamicTableUserInterface {
    [key: string]: StDynamicTableUISpecification;
+}
+
+export interface StDynamicTableFk {
+   table: string;
+   field: string;
 }