+++ projects/egeo/src/lib/st-dynamic-table/utils/st-dynamic-table.utils.ts
@@ -32,6 +32,8 @@ export class StDynamicTableUtils {
                      id: _propertyKey,
                      label: _property.title || _propertyKey,
                      reference: _property.$ref,
+                     fk: _property.fk,
+                     group: _property.group_field ? _property.group_field.name : null,
                      filters: filters,
                      filterable: filters && (filters.templateRef !== undefined || (filters.filterConfig && filters.filterConfig.length > 0)),
                      sortable: isSortable